+++ internal/graphapi/organization_test.go
@@ -10,6 +10,7 @@ import (
 	"github.com/brianvoe/gofakeit/v7"
 	"github.com/samber/lo"
 	"github.com/theopenlane/entx"
+	"github.com/theopenlane/entx/history"
 	"github.com/theopenlane/utils/ulids"
 	"gotest.tools/v3/assert"
 	is "gotest.tools/v3/assert/cmp"
@@ -20,7 +21,14 @@ import (
 	"github.com/theopenlane/core/common/models"
 	"github.com/theopenlane/core/internal/ent/generated"
 	ent "github.com/theopenlane/core/internal/ent/generated"
+	"github.com/theopenlane/core/internal/ent/generated/file"
+	"github.com/theopenlane/core/internal/ent/generated/organization"
 	"github.com/theopenlane/core/internal/ent/generated/privacy"
+	"github.com/theopenlane/core/internal/ent/generated/task"
+	"github.com/theopenlane/core/internal/ent/generated/trustcenter"
+	"github.com/theopenlane/core/internal/ent/generated/trustcentersetting"
+	"github.com/theopenlane/core/internal/ent/historygenerated/filehistory"
+	"github.com/theopenlane/core/internal/ent/historygenerated/taskhistory"
 	"github.com/theopenlane/core/internal/ent/privacy/rule"
 	"github.com/theopenlane/core/internal/graphapi/common"
 	"github.com/theopenlane/core/internal/graphapi/testclient"
@@ -1020,6 +1028,35 @@ func TestMutationOrganizationCascadeDelete(t *testing.T) {
 	// add child org
 	childOrg := (&OrganizationBuilder{client: suite.client, ParentOrgID: org.ID}).MustNew(reqCtx, t)
 
+	// a task gives us an org owned record that tracks history
+	task1 := (&TaskBuilder{client: suite.client}).MustNew(reqCtx, t)
+
+	allowCtx := setContext(reqCtx, suite.client.db)
+
+	// the trust center is org owned, but the setting created alongside it is not, it only points at
+	// the trust center. The cascade has to recurse through the trust center to reach it
+	trustCenter := (&TrustCenterBuilder{client: suite.client}).MustNew(reqCtx, t)
+
+	trustCenterSetting, err := suite.client.db.TrustCenterSetting.Query().
+		Where(trustcentersetting.TrustCenterID(trustCenter.ID)).First(allowCtx)
+	assert.NilError(t, err)
+
+	// the storage path is what the file hook hands to the object storage provider on delete
+	storageKey := "organizations/" + org.ID + "/cascade-test-object"
+
+	file1, err := suite.client.db.File.Create().
+		SetProvidedFileName("cascade-test.txt").
+		SetProvidedFileExtension("txt").
+		SetDetectedContentType("text/plain").
+		SetStoragePath(storageKey).
+		SetURI("file:///tmp/cascade-test.txt").
+		AddOrganizationIDs(org.ID).
+		Save(allowCtx)
+	assert.NilError(t, err)
+
+	// the history rows have to exist up front, otherwise asserting they are gone proves nothing
+	assertHistoryExists(t, allowCtx, task1.ID, file1.ID, true)
+
 	// delete org
 	resp, err := suite.client.api.DeleteOrganization(reqCtx, org.ID)
 
@@ -1051,12 +1088,58 @@ func TestMutationOrganizationCascadeDelete(t *testing.T) {
 		return generated.IsNotFound(err)
 	}, "custom domain should be deleted by async edge cleanup")
 
-	// verify the parent org is soft-deleted (not hard-deleted) by querying the db directly
-	ctx := privacy.DecisionContext(reqCtx, privacy.Allow)
-	ctx = entx.SkipSoftDelete(ctx)
+	// skipping soft delete makes soft deleted rows visible, so the assertions below fail if the
+	// cascade only marked the records deleted instead of removing them
+	purgedCtx := entx.SkipSoftDelete(privacy.DecisionContext(reqCtx, privacy.Allow))
 
-	o, err := suite.client.db.Organization.Get(ctx, org.ID)
+	// the organization row is removed last, once everything it owned is gone, so its absence is
+	// what tells us the whole cascade finished. Waiting for the queue to go idle is not enough on
+	// its own, a handler that errors parks the job in a retryable state the idle check ignores
+	waitForCondition(t, func() bool {
+		exists, err := suite.client.db.Organization.Query().Where(organization.ID(org.ID)).Exist(purgedCtx)
+
+		return err == nil && !exists
+	}, "organization should be hard deleted by async edge cleanup")
+
+	taskExists, err := suite.client.db.Task.Query().Where(task.ID(task1.ID)).Exist(purgedCtx)
 	assert.NilError(t, err)
-	assert.Equal(t, o.ID, org.ID)
-	assert.Assert(t, !o.DeletedAt.IsZero())
+	assert.Check(t, !taskExists, "task should be hard deleted with the organization, not soft deleted")
+
+	fileExists, err := suite.client.db.File.Query().Where(file.ID(file1.ID)).Exist(purgedCtx)
+	assert.NilError(t, err)
+	assert.Check(t, !fileExists, "file should be hard deleted with the organization, not soft deleted")
+
+	trustCenterExists, err := suite.client.db.TrustCenter.Query().Where(trustcenter.ID(trustCenter.ID)).Exist(purgedCtx)
+	assert.NilError(t, err)
+	assert.Check(t, !trustCenterExists, "trust center should be hard deleted with the organization")
+
+	// the setting carries no organization id of its own, it is only reachable by recursing through
+	// the trust center, so this is what proves the nested cleanup runs
+	settingExists, err := suite.client.db.TrustCenterSetting.Query().
+		Where(trustcentersetting.ID(trustCenterSetting.ID)).Exist(purgedCtx)
+	assert.NilError(t, err)
+	assert.Check(t, !settingExists, "trust center setting should be hard deleted with the organization")
+
+	assertHistoryExists(t, purgedCtx, task1.ID, file1.ID, false)
+
+	// the file hook only reaches object storage on a hard delete, a soft delete leaves the object
+	assert.Check(t, suite.client.deletedStorageKeys.Has(storageKey),
+		"the object backing the deleted file should have been removed from object storage")
+}
+
+// assertHistoryExists checks whether the history rows for the given task and file are present
+func assertHistoryExists(t *testing.T, ctx context.Context, taskID, fileID string, want bool) {
+	t.Helper()
+
+	historyCtx := history.WithContext(ctx)
+
+	taskHistory, err := suite.client.db.HistoryClient.TaskHistory.Query().
+		Where(taskhistory.Ref(taskID)).Exist(historyCtx)
+	assert.NilError(t, err)
+	assert.Check(t, is.Equal(want, taskHistory), "task history presence mismatch")
+
+	fileHistory, err := suite.client.db.HistoryClient.FileHistory.Query().
+		Where(filehistory.Ref(fileID)).Exist(historyCtx)
+	assert.NilError(t, err)
+	assert.Check(t, is.Equal(want, fileHistory), "file history presence mismatch")
 }
